+++ game_patch/misc/misc.cpp
@@ -443,6 +443,7 @@ CallHook level_init_pre_console_output_hook{
                         case rf::NetGameType::NG_TYPE_DC:
                             server_info.dc_score_limit = g_local_pending_win_condition.value();
                             break;
+                        case rf::NetGameType::NG_TYPE_ESC:
                         case rf::NetGameType::NG_TYPE_REV:
                             break;
                         case rf::NetGameType::NG_TYPE_RUN: